Finalize shooting gallery

+++ Assets/Scripts/PlayerStatistics.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class PlayerStatistics
 {
     public int Tickets;
-    public float SkeeballHighScore;
-    public float PlinkoHighScore;
-    public float ShootingGalleryHighScore;
+    public int SkeeballHighScore;
+    public int PlinkoHighScore;
+    public int ShootingGalleryHighScore;
 
-    public float RingTossHighScore;
+    public int RingTossHighScore;
 }
